Needs wrapping.
git-svn-id: svn://falconpl.org/falcon/modules/sdl/trunk@100 08e93009-9248-0410-8e9b-f56be5c5ec23

+++ ttf/sdlttf_ext.cpp
@@ -222,7 +222,7 @@ FALCON_FUNC ttf_OpenFont( VMachine *vm )
       return;
    }
 
-   long index = i_index != 0 ? i_index->forceInteger() : 0;
+   long index = i_index != 0 ? ((long)i_index->forceInteger()) : 0;
    AutoCString file( *i_filename->asString() );
 
    ::TTF_Font *fnt = ::TTF_OpenFontIndex( file.c_str(),
